Made tag text a primary key, improved mood update layout, made tag rating average a float
git-svn-id: https://subversion.assembla.com/svn/ThoughtFerret@80 578f6600-eed7-407d-a15d-0bb1fb5aace9

+++ trunk/source/ThoughtFerret/src/com/thoughtworks/thoughtferret/model/tags/MoodTag.java
@@ -25,8 +25,8 @@ public class MoodTag {
 		return ratingSum;
 	}
 
-	public int getRatingAverage() {
-		return ratingSum / count;
+	public float getRatingAverage() {
+		return ratingSum / (float)count;
 	}
 	
 	public MoodTag add(MoodTag other) {
